# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: SML Staff Marketplace API
  version: "1.1.0"
  description: |
    Express.js port of the original Java/JAX-RS MarketPlaceWebService.

    All endpoints are under `/service/v1` (included in the server URL below).

    **Auth note**: Failed login returns HTTP 200 with `data: []` — not HTTP 401.

    **Image endpoints**: Return binary `image/png`. `/imagesguid` never returns 404 — returns empty buffer instead.
servers:
  - url: http://localhost:47309/service/v1
    description: Local development (current SML Staff Ubon)
  - url: http://localhost:47300/service/v1
    description: Local development (legacy preview)
...
security:
  - GUIDHeader: []
components:
  securitySchemes:
    GUIDHeader:
      type: apiKey
      in: header
      name: GUID
      description: Session identifier required by most endpoints
...
paths:
...
  /purchase-print/render:
    get:
      tags: [Purchase, Print]
      summary: Render ฟอร์มพิมพ์ PU เป็น HTML
      description: |
        Returns browser-printable HTML using the shared sale print renderer.

        If `formcodes` is omitted, the first available PU form is used. If `log_print` is omitted,
        print logging follows `auto_print` (`auto_print=0` previews without logging by default).
      security: []
      parameters:
        - { name: doc_no, in: query, required: true, schema: { type: string }, example: "PU-26050018" }
        - { name: formcodes, in: query, schema: { type: string }, example: "PUPOR1" }
        - { name: auto_print, in: query, schema: { type: string, enum: ["0", "1"], default: "1" }, description: '"1" injects window.print behavior in HTML' }
        - { name: log_print, in: query, schema: { type: string, enum: ["0", "1"] }, description: "Explicit print log override; use 0 for preview" }
        - { name: user_code, in: query, schema: { type: string }, example: "001" }
      responses:
        '200':
          description: Rendered HTML
          headers:
            Cache-Control:
              schema: { type: string, example: "no-store" }
          content:
            text/html:
              schema: { type: string }
        '400':
          description: doc_no is required
          content:
            text/plain:
              schema: { type: string }
        '404':
          description: document or print form not found
          content:
            text/plain:
              schema: { type: string }
        '500':
          description: Render error
          content:
            text/plain:
              schema: { type: string }
...